feat: refactor cache handling, move storage out of generated (#238)
* feat: move introspection cache to .wundergraph/introspection/cache

This removes this from .wundergraph/generated, so it can be checked into
the VCS

* feat: prioritize cache during introspection

Always try to hit the cache first. Once a result is retrieved, populate
the cache as long as it's not disabled. Make the cache enabled by
default, but toggleable via WG_DISABLE_INTROSPECTION_CACHE.

* fix: ammend cache path in the SDK side

* fix: print a log when writing to cache fails

* chore: move introspection cache to $WGD/cache/introspection

This way we can eventually place more items into a cache using the
hierarchy below $WGD/cache

* chore: handle introspection cache dir creation in the TS side

wunderctl generate was not propertly initializing it. Instead of making
commands that might end up using the cache initialize the directory,
make the TS side initialize it lazily when needed.

* fix: fix race condition in readIntrospectionCacheFile()

Instead of testing for file existence, try to read it and handle the
potential errors.

* feat: add global --cache flag to control the cache

- New --cache flag enables caching, which otherwise is always disabled
  by default
- --cache is implemented globally for all commands, although it doesn't
  have any effect for some of them
- TS side honors this via WG_ENABLE_INTROSPECTION_CACHE environment
  variable

* feat: improve cache UI

- Enable cache by default
- During wunderctl generate, fail if there's a cache miss
- Add --cache-fallback flag to wunderctl generate, which allows falling
  back onto introspection
- Pass WG_ENABLE_INTROSPECTION_CACHE explicitely instead of setting it
  globally
- Clear the cache by default during wunderctl up startup
- Add a configurable time interval to clear the cache during wunderctl
  up, with a 5 seconds default

* fix: don't clear the cache from the go side

Let the TS side do the refreshes when polling is enabled

* fix: fix help for --no-cache flag

* fix: remove --keep-introspection-cache flag from wunderctl up

Always clear it up during startup

* chore: rename --cache-fallback logic to use --offline, flip the logic

The UI ends up much more intuitive this way

* docs: update docs for cache handling

* chore: update help texts for wunderctl generate

* docs: reword help for --no-cache and --ofline flags

* feat: improve error message in offline mode

Make it clear that the error is because offline mode is enabled, include
details of the introspection that failed

* feat: add global flag for clearing caches

As discussed with Dustin, this is useful for generating a cache from a
known and cold state, to reuse it later (e.g. for CI runs)

* chore: add .wundergraph/cache to ignore files

Added to both .gitignore and .prettierignore, as well as testsapps and
examples

* docs: add examples for --clear-cache and --offline

* chore: add .wundergraph/cache to new .gitignore example files