feat: add configurable per-source timeouts (#232)
* feat: add configurable per-source timeouts

- Add a new optional timeout field in HTTPUpstream
- Pass this new field to the node via FetchConfiguration, so it applies
  to all HTTP based sources
- In the server, create a dedicated http.Client with the specified
  timeout as needed

* feat: add support for timeouts to database based sources

Move timeout parameter to DataSourceConfiguration, so all sources can
now have a timeout regardless of their type. At some point, we might
want to implement timeouts for all of them.

* fix: edit pb.go file to match protoc version used in CI

* feat: add default timeout for the whole app

Add new optional field in NodeOptions with the timeout in milliseconds.
For now, it only applies to HTTP based requests.

* fix: rename timeout: to timeoutMilliseconds: in SDK

Having the unit in the field name makes it more evident

* feat: allow setting unlimited timeouts

null-ish or < 0 will use the default timeout, but 0 will make it
unlimited

* fix: remove leftover field from HTTPUpstream

timeout is not used anymore, now it applies to all data sources

* fix: api changes

timeoutMilliseconds -> requestTimeoutMilliseconds

Also, disallow disabling the timeout by setting it to zero. Zero keeps
the default value instead.

* chore: use seconds instead of milliseconds for request timeouts

* fix: fix newly added test

* docs: fix incorrect field name in introspect.graphql() example

* fix: override graphql_datasource.Factory HTTPClient

Instead of creating a new instance, overwrite the HTTPClient field in
the existing one

* refactor: useCustomHTTPClient() -> requiresCustomHTTPClient()

* docs: document added field to exported types

* fix: update generated data in snapshot used for testing