feat(create-wundergraph-app): download latest tag instead of main (#349)
* feat(create-wundergraph-app): download latest tag instead of main

Downloading from the main branch prevents us from making breaking
changes, since the examples would stop working until the next SDK with
the breaking changes is released.

When using the wundegraph repository as the example source, first
retrieve the tags and use the latest tag starting with @wundergraph/sdk

* fix: download from actual ref instead of 'main'

When an example was provided in the command line, the ref was incorrectly
set to 'main'

* fix: path filtering when branch/tag contains path-invalid characters

When the ref name contains characters that are not valid in a path,
the top directory name doesn't match the ref name. Instead, strip
the first level and match against the rest.