feat: add WunderGraph testing library (#345)
* feat: add wunderctl node url subcommand

Prints the node URL

* feat: add wunderctlSubprocess

Returns a process that controls a wunderctl instance

* feat: add initial testing framework

Automatically spins up and tears down wunderctl during tests

* feat: add tests to simple example

* feat: add support for mocking origins in tests

- Add pkg/httpinterceptor for intercepting http.RoundTripper
- Add support in pkg/node and pkg/apihandler for wiring up an interceptor
- Add up command --intercept-http argument that sends requests to an
intermediate server for interception
- Start a web server while running tests, which intercepts HTTP requests
coming from the node
- Implement minimal compatibility classes for Headers, Request and Response,
to use them in origin mocks

* feat: add Server.runTest() shorthand

Wraps a test and runs it

* feat: add Server.clientConfig()

Returns a Client configuration which allows intercepting requests

* feat: add support for returning error messages in mocks

Eventually used to disable connecting to the network during tests

* feat: add Server.client()

Returns a TS client configured for testing

* chore: remove HTTP interception and mocking support

Removing while we settle on the final API

* chore: use node start instead of up when running tests

Makes spinup faster by avoiding introspection step

* chore: spin and tear down server only once per block

Instead of shutting down the node inmmediately when a test ends,
wait a bit to see if another one starts.

* chore: add support for passing environment vars to wunderctl

* feat: add support for parallel testing.Server instances

Use random ports for the node and the hooks server, so we can run
several instances at the same time. There's a small chance of a race
condition since we're picking the port in the TS side and using it in
the Go side, but it shouldn't be a problem in practical terms.

* chore: remove wunderctl node url

Since we're now generating the ports/urls in the TS side, we no longer
need to retrieve the default one

* feat: add support for manually running test server, debugging

- Add support for manually starting the server to run the test against
manually, by setting WG_NODE_URL before running the test. This gives users
a escape hatch to setup their own node if they need to
- Add a server option to enable debug logging, which can be turned also
by the WG_TEST_DEBUG environment variable

* feat: add tests for openapi-mocking example using ava

* docs: add docs for testing library

* refactor: spinUp() => setUp()

* docs: fix missing ; in testing document

* chore: remove (now) unused TestOptions type

* fix(examples): better name for main test section in simple

* fix(testing): check the actual server health before continuing

We were only checking for the status code

* fix(testing): swallow error when shutting down the test server takes too long

* refactor: rename ClientConfigInit to CreateClientConfig

* docs: rephrase mention of our TypeScript client

* docs: rephrase some concepts in the testing library docs

* feat: remove automatic API, rename methods to start/stop

Update docs to reflect the change

* docs: headless -> integration

* refactor: testing.Server => testing.WunderGraphTestServer

* feat: add generated template for testing library

This allows users to just import the generated testing file, without
having to manually import the TS client and pass it to the server.
Update docs accordingly

* fix: specify node-fetch type as a dev dependency

* fix: add semicolons in docs

* fix: clear request timeout when checking health endpoint

It was not cleared correctly in case fetch failed (e.g. connection
refused)

* refactor: use ts-retry-promise to implement health check retry logic

Simplifies the logic and lets handle the startup timeout implicitly,
without having to count the actual elapsed time.

* refactor: remove public API for timeout, default to 10 retries

Each retry is limited to a maximum of 5 seconds. If someone complains,
we'll add an API to expose it.

Co-authored-by: Dustin Deus <deusdustin@gmail.com>