fix: data corruption when polling is enabled for federation sources (#753)
* fix: data corruption when polling is enabled for federation sources

Federation was querying its graphql components by using an introspection
function that goes through the cache, which in turn launched an introspection
in an interval when polling was enabled. Since federation called each of its
graphql sources when polling, this resulted in a new scheduled loop on
every call.

To fix this, expose a graphql introspection function that doesn't use the
cache and use it for introspecting federated sources. This shouldn't be
a problem because we do cache the result of the federation.

Also, add a few debug messages and handle a few more error cases when
introspecting graphql (as a result of the debugging journey).

Additioanlly, there was also a race condition in the cache files, which
caused problems with concurrent writes. To fix this, use atomic writes
when creating the cache files.