fix: print compiler error messages during TS operation response schema generation (#761)
Previously, we were silencing error messages from typescript-json-schema because
we have some fallback logic to handle some errors. This is a problem when
we get an error that cannot be recovered because we were not printing what the
error actually was.

Instead, we now first determine if we can find a tsconfig.json and in that case, we call typescript-json-schema
using it. Otherwise, we call it with default arguments. In both cases, we call it only once so we don't need to
silence the errors and they'll get printed to the console.