feat: add support for server ping during subscription idle times (#1234)
Some providers or middleboxes aggressively close TCP connections that
look inactive. This means a long-running subscription might be 
disconnected if no new data arrives for some time. To counteract this,
introduce a ping mechanism in the server that ensures that the line is
not idle for longer than the requested interval by sending an empty
payload if no actual data has been sent. This is enabled by default
with an interval of 30s and can be overriden by setting the
WG_SUBSCRIPTION_SERVER_PING_INTERVAL environment variable.

Also, refactor all the SSE / jsonpatch support to implement it in a single
place.

Additionally, fix a bug that incorrectly shared a variable in the client
when returning subscription results to the caller.